Move terminal and balance of in Server On Pages folder
