+++ extraResources/config.json
@@ -1,1 +1,1 @@
-{"NAME":"ewqewqewqe","PAIRS":{"DAI-ETH":{"FEE":"3"}},"WALLETS":{"ETH":{"ADDRESS":"0xfCd50a1620f21A46Ddf39B526DCA17c59D94e323","SECRET":"F877C159A076F27D989E506F11AB4FADBE6694F9857E43107F049B4D9901A6A0"},"DAI":{"ADDRESS":"0xfCd50a1620f21A46Ddf39B526DCA17c59D94e323","SECRET":"F877C159A076F27D989E506F11AB4FADBE6694F9857E43107F049B4D9901A6A0"}},"PRICE":{"PROVIDER":"CryptoCompare","API_KEY":"32131211","INTERVAL":30},"NOTIFICATIONS":{},"AGGREGATOR_URL":"https://network.jelly.market/api/v1/info/update","SERVER":{"PORT":"9000"},"DATABASE":{"ACTIVE":"SQLITE","SQLITE":{"database":"butler.sqlite"}}}
+{"NAME":"ewqewqewqe","PAIRS":{"DAI-ETH":{"FEE":"3"}},"WALLETS":{"ETH":{"ADDRESS":"0xfCd50a1620f21A46Dde39B526DCA17c59D94e323","SECRET":"F877C159A076F27D989E506F11AB4FADBE6694F9857E43107F049B4D9901A6A0"},"DAI":{"ADDRESS":"0xfCd50a1620f22A46Ddf39B526DCA17c59D94e323","SECRET":"F877E159A076F27D989E506F11AB4FADBE6694F9857E43107F049B4D9901A6A0"}},"PRICE":{"PROVIDER":"CryptoCompare","API_KEY":"32131211","INTERVAL":30},"NOTIFICATIONS":{},"AGGREGATOR_URL":"https://network.jelly.market/api/v1/info/update","SERVER":{"PORT":"9000"},"DATABASE":{"ACTIVE":"SQLITE","SQLITE":{"database":"butler.sqlite"}}}
