+++ src/components/pages/serverONPages/balanceOf/style.scss
@@ -1,4 +1,4 @@
-@import '../../../css/constants.scss';
+@import "../../../../css/constants.scss";
 
 .balance-of-wrapper {
   background-color: $bgColor;
