+++ src/components/pages/serverONPages/terminal/index.js
