+++ src/components/pages/serverONPages/terminal/style.scss
