+++ src/components/router.js
@@ -2,8 +2,6 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import Questions from './pages/questions';
-import Terminal from './pages/terminal';
-import BalanceOf from './pages/balanceOf';
 import ServerONPages from './pages/serverONPages';
 
 const ReactRouter = () => (
